+ Added simple try count handling and JMX metrics of client queue.

+++ http-adapter/src/main/java/com/a1systems/http/adapter/message/MessagePart.java
@@ -156,7 +156,7 @@ public class MessagePart implements Delayed{
         this.state = state;
     }
 
-    public int getTryCount() {
+    public byte getTryCount() {
         return tryCount;
     }
